# Projeto2026/api_dados/swagger.yaml
openapi: 3.0.3
info:
  title: EW API de Dados
  version: 1.0.0
  description: API de dados para produtos, encomendas, cores e categorias.
servers:
  - url: http://localhost:7789
    description: Servidor local de desenvolvimento
...
paths:
  /produtos:
    get:
      tags: [Produtos]
      summary: Listar produtos
      description: |
        Lista produtos com filtros opcionais. Os filtros suportados incluem categoria, cor, tamanho,
        intervalo de preço, pesquisa por nome, ordenação e paginação.
      security: []
      parameters:
        - in: query
          name: categoria
          schema:
            type: string
          description: Identificador da categoria. O servidor normaliza o valor para formato com hífens.
        - in: query
          name: cor
          schema:
            type: string
          description: Identificador da cor em hexadecimal armazenado em `opcoes.cores._id`.
        - in: query
          name: tamanho
          schema:
            type: string
          description: Tamanho a procurar em `opcoes.tamanhos`.
        - in: query
          name: minPreco
          schema:
            type: number
          description: Preço mínimo para filtrar variantes.
        - in: query
          name: maxPreco
          schema:
            type: number
          description: Preço máximo para filtrar variantes.
        - in: query
          name: sort
          schema:
            type: string
            enum: [preco_asc, preco_desc, nome_asc, nome_desc]
          description: Ordenação por preço ou nome.
        - in: query
          name: search
          schema:
            type: string
          description: Pesquisa parcial no nome do produto.
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
          description: Página da paginação.
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
          description: Número de registos por página.
      responses:
        '200':
          description: Lista de produtos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProdutosListResponse'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [Produtos]
      summary: Criar produto
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [produto]
              properties:
                produto:
                  type: string
                  description: JSON serializado com os dados do produto.
      responses:
        '201':
          description: Produto criado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /produtos/stats:
    get:
      tags: [Produtos]
      summary: Obter estatísticas de produtos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Estatísticas agregadas dos produtos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProdutoStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /encomendas:
    get:
      tags: [Encomendas]
      summary: Listar encomendas
      description: |
        Requer autenticação. Clientes só conseguem ver as próprias encomendas; administradores podem consultar todas.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: userId
          schema:
            type: string
          description: Filtra pelo utilizador. Para clientes, o servidor ignora este valor e usa o ID do token.
        - in: query
          name: estado
          schema:
            type: string
            enum: [pendente, concluida, cancelada]
        - in: query
          name: dataInicio
          schema:
            type: string
            format: date
          description: Data inicial do intervalo.
        - in: query
          name: dataFim
          schema:
            type: string
            format: date
          description: Data final do intervalo.
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
        - in: query
          name: sort
          schema:
            type: string
          description: Campo de ordenação passado diretamente para o Mongoose.
      responses:
        '200':
          description: Lista de encomendas.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EncomendasListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [Encomendas]
      summary: Criar encomenda
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EncomendaInput'
      responses:
        '201':
          description: Encomenda criada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Encomenda'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /encomendas/stats:
    get:
      tags: [Encomendas]
      summary: Obter estatísticas de encomendas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Estatísticas por estado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EncomendaStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /cores:
    get:
      tags: [Cores]
      summary: Listar cores
      security: []
      responses:
        '200':
          description: Lista de cores.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cor'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [Cores]
      summary: Criar cor
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CorInput'
      responses:
        '201':
          description: Cor criada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /categorias:
    get:
      tags: [Categorias]
      summary: Listar categorias
      security: []
      responses:
        '200':
          description: Lista de categorias.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Categoria'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [Categorias]
      summary: Criar categoria
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoriaInput'
      responses:
        '201':
          description: Categoria criada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Categoria'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        O middleware aceita o token no header `Authorization: Bearer <token>`, cookie `token` ou query string `?token=`.
...
  responses:
    BadRequest:
      description: Pedido inválido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Não autenticado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthErrorResponse'
    Forbidden:
      description: Sem permissões para executar a operação.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthErrorResponse'
...
    ServerError:
      description: Erro interno do servidor.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Produto not found
        erro:
          type: string
          example: Token inválido ou expirado.
    AuthErrorResponse:
      type: object
      properties:
        erro:
          type: string
      example:
        erro: Acesso negado. Requer permissão de administrador.
...
    ProdutoColorOption:
      type: object
      properties:
        _id:
          type: string
          description: Hex da cor.
        label:
          type: string
        imagem:
          type: string
...
    ProdutoBase:
      type: object
      properties:
        _id:
          type: string
        nome:
          type: string
        categoriaId:
          type: string
        imagemBase:
          type: string
        opcoes:
          type: object
          properties:
            cores:
              type: array
              items:
                $ref: '#/components/schemas/ProdutoColorOption'
            tamanhos:
              type: array
              items:
                type: string
            tampas:
              type: array
              items:
                type: string
        variantes:
          type: array
          items:
            $ref: '#/components/schemas/ProdutoVariant'
...
    Produto:
      allOf:
        - $ref: '#/components/schemas/ProdutoBase'
    ProdutosListResponse:
      type: object
      properties:
        produtos:
          type: array
          items:
            $ref: '#/components/schemas/Produto'
        total:
          type: integer
    ProdutoStats:
      type: object
      properties:
        numProdutos:
          type: integer
        vendas:
          type: integer
        precoMedio:
          type: string
          example: '12.50'
        faturacao:
          type: string
          example: '1450.00'
...
    EncomendaInput:
      allOf:
        - $ref: '#/components/schemas/EncomendaBase'
      required:
        - userId
        - produtos
        - precoTotal
...
    Encomenda:
      allOf:
        - $ref: '#/components/schemas/EncomendaBase'
    EncomendasListResponse:
      type: object
      properties:
        encomendas:
          type: array
          items:
            $ref: '#/components/schemas/Encomenda'
        total:
          type: integer
    EncomendaStats:
      type: object
      properties:
        concluidas:
          type: integer
        pendentes:
          type: integer
        canceladas:
          type: integer
        total:
          type: integer
    CorInput:
      type: object
      properties:
        _id:
          type: string
          description: Hex da cor.
        label:
          type: string
      required:
        - label
    Cor:
      type: object
      properties:
        _id:
          type: string
        label:
          type: string
    CategoriaInput:
      type: object
      properties:
        _id:
          type: string
        label:
          type: string
      required:
        - label
    Categoria:
      type: object
      properties:
        _id:
          type: string
        label:
          type: string
...